more protocol stuff (r1.8 -> r1.7)

+++ src/main/java/com/github/dirtpowered/dirtmv/network/versions/Beta17To14/ProtocolBeta17to14.java
@@ -102,7 +102,7 @@ public class ProtocolBeta17to14 extends ServerProtocol {
 
                 return PacketUtil.createPacket(0x08, new TypeHolder[]{
                         data.read(0),
-                        set(Type.SHORT, 6),
+                        set(Type.SHORT, (short) 6),
                         set(Type.FLOAT, 0.0F),
 
                 });
